+++ actions/setup-telemetry/CHANGELOG.md
@@ -1,5 +1,22 @@
 # Install Setup Telemetry
 
+## 0.1.5
+
+### Patch Changes
+
+- 37cb5e9: Upgrade OpenTelemetry dependencies
+- Updated dependencies [37cb5e9]
+  - @pagopa/azure-tracing@0.4.14
+
+## 0.1.4
+
+### Patch Changes
+
+- d634c90: Remove the `@azure/monitor-opentelemetry` and add `@pagopa/azure-tracing` dependency.
+
+  By doing this, every time we update the internal `@pagopa/azure-tracing` package, we are propagating the
+  changes to the `setup-telemetry` action as well.
+
 ## 0.1.3
 
 ### Patch Changes