+++ infra/modules/azure_storage_account/CHANGELOG.md
@@ -1,5 +1,11 @@
 # azure_storage_account
 
+## 2.2.3
+
+### Patch Changes
+
+- 13b31e8: Suppress Trivy warnings
+
 ## 2.2.2
 
 ### Patch Changes